Animations playing for a SkinnedMeshComponent.
Transforms are off (look into this).

+++ Resources/Shaders/hlsl/Common_Structs.hlsl
@@ -6,7 +6,7 @@ struct GeoVertexInput
 	float3 Colour : COLOR0;
 	float2 UV : TEXCOORD0;
 
-	float4 BoneIds : TEXCOORD1;
+	uint4 BoneIds : TEXCOORD1;
 	float4 BoneWeights : TEXCOORD2;
 };
